Compute the metrics in the lambda invocation
